Fixes 12 lines of compiler warnings for me, and doesn't seem to affect the code much.
svn path=/trunk/kdebase/kioslave/; revision=231496

+++ nfs/mount_xdr.c
@@ -46,7 +46,7 @@ bool_t
 xdr_fhandle(XDR *xdrs, fhandle objp)
 {
 
-	 register long *buf=buf;
+	 register int32_t *buf=buf;
 
 	 if (!xdr_opaque(xdrs, objp, FHSIZE)) {
 		 return (FALSE);
@@ -58,7 +58,7 @@ bool_t
 xdr_fhstatus(XDR *xdrs, fhstatus *objp)
 {
 
-	 register long *buf=buf;
+	 register int32_t *buf=buf;
 
 	 if (!xdr_u_int(xdrs, &objp->fhs_status)) {
 		 return (FALSE);
@@ -79,7 +79,7 @@ bool_t
 xdr_dirpath(XDR *xdrs, dirpath *objp)
 {
 
-	 register long *buf=buf;
+	 register int32_t *buf=buf;
 
 	 if (!xdr_string(xdrs, objp, MNTPATHLEN)) {
 		 return (FALSE);
@@ -91,7 +91,7 @@ bool_t
 xdr_name(XDR *xdrs, name *objp)
 {
 
-	 register long *buf=buf;
+	 register int32_t *buf=buf;
 
 	 if (!xdr_string(xdrs, objp, MNTNAMLEN)) {
 		 return (FALSE);
@@ -103,7 +103,7 @@ bool_t
 xdr_mountlist(XDR *xdrs, mountlist *objp)
 {
 
-	 register long *buf=buf;
+	 register int32_t *buf=buf;
 
 	 if (!xdr_pointer(xdrs, (char **)objp, sizeof(struct mountbody), (xdrproc_t)xdr_mountbody)) {
 		 return (FALSE);
@@ -115,7 +115,7 @@ bool_t
 xdr_mountbody(XDR *xdrs, mountbody *objp)
 {
 
-	 register long *buf=buf;
+	 register int32_t *buf=buf;
 
 	 if (!xdr_name(xdrs, &objp->ml_hostname)) {
 		 return (FALSE);
@@ -133,7 +133,7 @@ bool_t
 xdr_groups(XDR *xdrs, groups *objp)
 {
 
-	 register long *buf=buf;
+	 register int32_t *buf=buf;
 
 	 if (!xdr_pointer(xdrs, (char **)objp, sizeof(struct groupnode), (xdrproc_t)xdr_groupnode)) {
 		 return (FALSE);
@@ -145,7 +145,7 @@ bool_t
 xdr_groupnode(XDR *xdrs, groupnode *objp)
 {
 
-	 register long *buf=buf;
+	 register int32_t *buf=buf;
 
 	 if (!xdr_name(xdrs, &objp->gr_name)) {
 		 return (FALSE);
@@ -160,7 +160,7 @@ bool_t
 xdr_exports(XDR *xdrs, exports *objp)
 {
 
-	 register long *buf=buf;
+	 register int32_t *buf=buf;
 
 	 if (!xdr_pointer(xdrs, (char **)objp, sizeof(struct exportnode), (xdrproc_t)xdr_exportnode)) {
 		 return (FALSE);
@@ -172,7 +172,7 @@ bool_t
 xdr_exportnode(XDR *xdrs, exportnode *objp)
 {
 
-	 register long *buf=buf;
+	 register int32_t *buf=buf;
 
 	 if (!xdr_dirpath(xdrs, &objp->ex_dir)) {
 		 return (FALSE);
@@ -190,12 +190,12 @@ bool_t
 xdr_ppathcnf(XDR *xdrs, ppathcnf *objp)
 {
 
-	 register long *buf=buf;
+	 register int32_t *buf=buf;
 
 	 int i=i;
 
 	 if (xdrs->x_op == XDR_ENCODE) {
-	 buf = XDR_INLINE(xdrs,6 * BYTES_PER_XDR_UNIT);
+	 buf = XDR_INLINE(xdrs, 6 * BYTES_PER_XDR_UNIT);
 	   if (buf == NULL) {
 		 if (!xdr_int(xdrs, &objp->pc_link_max)) {
 			 return (FALSE);
@@ -285,7 +285,7 @@ xdr_ppathcnf(XDR *xdrs, ppathcnf *objp)
 	 if (!xdr_char(xdrs, &objp->pc_xxx)) {
 		 return (FALSE);
 	 }
-		buf = XDR_INLINE(xdrs,   2  * BYTES_PER_XDR_UNIT);
+		buf = XDR_INLINE(xdrs, 2  * BYTES_PER_XDR_UNIT);
 		if (buf == NULL) {
 		 if (!xdr_vector(xdrs, (char *)objp->pc_mask, 2, sizeof(short), (xdrproc_t)xdr_short)) {
 			 return (FALSE);
